split API and views (#16)
